requirements.txt
black
croniter
matplotlib
mypy
numpy
pandas
pylint
python-crontab
python-dotenv
pytz
structlog
tqdm
typer
types-croniter
types-pytz
urllib3
